+++ ies_tools/columntype.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class ColumnType(Enum):
+    number = 0
+    str = 1
+    calc = 2