+++ ies_tools/propertyaccess.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class PropertyAccess(Enum):
+    EP = 0
+    CP = 1
+    VP = 2
+    SP = 3
+    CT = 4